harmonize updating teams
